+++ src/components/Globe/components/Arcs/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Arcs';
