+++ src/components/Globe/components/RegionOverlay/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RegionOverlay';
