+++ src/components/Globe/components/ServerMarkers/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ServerMarkers';
